# Check if current key is greater than searched key