--- a/atlas.go
+++ b/atlas.go
@@ -42,7 +42,7 @@ func (n *Node) Data() (prev *Node, key string, value string) {
 
 // ValueByKey gets the value of key written in this Node or any of its ancestor.
 func (n *Node) ValueByKey(k string) (string, bool) {
-	if n.root == n {
+	if n.root == n || n.linkKey[0] > k {
 		return "", false
 	}
 	if n.linkKey[0] == k {
@@ -109,6 +109,9 @@ func (n *Node) Contains(sub *Node) bool {
 	}
 	if n.root == n {
 		return false
+	}
+	if n.linkKey[0] > sub.linkKey[0] {
+		return n.prev.Contains(sub.prev)
 	}
 	// TODO: https://github.com/mstoykov/atlas/issues/2
 	// apparently this is faster than if n.linkKey == sub.linkKey