This commit resolves #12