@mstoykov 2022-08-11
for posterity bench results after the latest changes to the benchmarks in `main` 
```
name                          old time/op    new time/op    delta
NodeContainsPositive-8          11.5ns ±10%    13.6ns ± 5%   +18.69%  (p=0.008 n=5+5)
NodeContainsNegative-8          7.43ns ± 5%    5.16ns ± 5%   -30.59%  (p=0.008 n=5+5)
NodeValueByKey/goodKeyLow-8     2.46ns ± 6%    6.16ns ±10%  +150.55%  (p=0.008 n=5+5)
NodeValueByKey/badKeyLow-8      5.33ns ± 4%    8.67ns ±16%   +62.64%  (p=0.008 n=5+5)
NodeValueByKey/goodKeyMid-8     2.76ns ± 5%    5.51ns ±13%   +99.77%  (p=0.008 n=5+5)
NodeValueByKey/badKeyMid-8      5.34ns ± 4%    6.58ns ± 3%   +23.20%  (p=0.008 n=5+5)
NodeValueByKey/goodKeyHigh-8    5.66ns ± 5%    1.54ns ± 4%   -72.79%  (p=0.008 n=5+5)
NodeValueByKey/badKeyHigh-8     4.81ns ± 3%    1.56ns ± 3%   -67.65%  (p=0.008 n=5+5)

name                          old alloc/op   new alloc/op   delta
NodeContainsPositive-8           0.00B          0.00B           ~     (all equal)
NodeContainsNegative-8           0.00B          0.00B           ~     (all equal)
NodeValueByKey/goodKeyLow-8      0.00B          0.00B           ~     (all equal)
NodeValueByKey/badKeyLow-8       0.00B          0.00B           ~     (all equal)
NodeValueByKey/goodKeyMid-8      0.00B          0.00B           ~     (all equal)
NodeValueByKey/badKeyMid-8       0.00B          0.00B           ~     (all equal)
NodeValueByKey/goodKeyHigh-8     0.00B          0.00B           ~     (all equal)
NodeValueByKey/badKeyHigh-8      0.00B          0.00B           ~     (all equal)

name                          old allocs/op  new allocs/op  delta
NodeContainsPositive-8            0.00           0.00           ~     (all equal)
NodeContainsNegative-8            0.00           0.00           ~     (all equal)
NodeValueByKey/goodKeyLow-8       0.00           0.00           ~     (all equal)
NodeValueByKey/badKeyLow-8        0.00           0.00           ~     (all equal)
NodeValueByKey/goodKeyMid-8       0.00           0.00           ~     (all equal)
NodeValueByKey/badKeyMid-8        0.00           0.00           ~     (all equal)
NodeValueByKey/goodKeyHigh-8      0.00           0.00           ~     (all equal)
NodeValueByKey/badKeyHigh-8       0.00           0.00           ~     (all equal)
```